Updated MusicObject for editor testing
- Music now plays from whatever scene you start from and continues correctly
- MusicObject prefab in each scene in case it gets destroyed somehow (unlikely)
