+++ Assets/Prefabs/MusicObject.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1ff13523424914af5b440d7e7d08b094
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
