+++ Assets/Scenes/Intro.unity
@@ -880,6 +880,67 @@ MonoBehaviour:
     y: 0
     width: 1
     height: 1
+--- !u!1001 &1443202351
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 5368311931405240150, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_Name
+      value: MusicObject
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_RootOrder
+      value: 4
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 213.36893
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 7.4224806
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 247.32948
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
 --- !u!1 &2014146100
 GameObject:
   m_ObjectHideFlags: 0
