+++ Assets/Scenes/Outro.unity
@@ -650,6 +650,67 @@ CanvasRenderer:
   m_PrefabAsset: {fileID: 0}
   m_GameObject: {fileID: 1269949359}
   m_CullTransparentMesh: 1
+--- !u!1001 &1275229001
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: 5368311931405240150, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_Name
+      value: MusicObject
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_RootOrder
+      value: 4
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 213.36893
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: 7.4224806
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: 247.32948
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 7820327485412135092, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents: []
+  m_SourcePrefab: {fileID: 100100000, guid: 1ff13523424914af5b440d7e7d08b094, type: 3}
 --- !u!1 &1409734143
 GameObject:
   m_ObjectHideFlags: 0
